ethereum address declarative validator
