+++ src/index.js
@@ -1,5 +1,7 @@
 import App from './app'
+import { registryCustomValidators } from 'utils/validator'
 
 require('events').EventEmitter.defaultMaxListeners = 0
+registryCustomValidators()
 
 App.start()
