+++ src/models/ProfileModel.js
@@ -1,6 +1,4 @@
 import { abstractModel } from './AbstractModel'
-import validator from '../components/forms/validator'
-import ErrorList from '../components/forms/ErrorList'
 
 class ProfileModel extends abstractModel({
   name: null,
@@ -30,18 +28,6 @@ export const validateRules = {
   name: 'required|min:3',
   email: 'email',
   company: 'min:3',
-}
-
-/**
- * @deprecated use declarative rules
- * @param values
- * @return {*}
- */
-export const validate = values => {
-  errors.name = ErrorList.toTranslate(validator.name(values.get('name')))
-  errors.email = ErrorList.toTranslate(validator.email(values.get('email'), false))
-  errors.company = ErrorList.toTranslate(validator.name(values.get('company'), false))
-  return errors
 }
 
 export default ProfileModel
